changed mileage from int to float in db

+++ db/migrate/20180504165209_create_workouts_table.rb
@@ -4,7 +4,7 @@ class CreateWorkoutsTable < ActiveRecord::Migration[5.2]
       t.string    :workout
       t.string    :comment
       t.integer   :duration
-      t.integer   :mileage
+      t.float     :mileage
       t.datetime  :date
       t.integer   :user_id
     end
